refactor: rename h2_client to h2client and h2_pool to h2pool
Apply consistent naming convention by removing underscores from variable
and module names: h2_client -> h2client, h2_pool -> h2pool
