+++ src/lib.rs
@@ -1,5 +1,5 @@
 pub mod executor;
-pub mod h2_client;
+pub mod h2client;
 pub mod http;
 pub mod provider;
 pub mod websocket;
